+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 Added:
 - CRN.equilibrium returns a model that can be fit to experimental data
+- ns.jupyter.ion and ioff allow to turn interactive output on or off
 
 Changed:
 - CRN.equilibrate now accepts 2D initial conditions with specialized CRN.params